# ddtrace/tracer: Improve implementation of payload