Hi,

There is a bug in the current implementation of tracer.payload, which leads to a potential mismatch between the payload's content (in payload.buf) and the payload trace count (payload.count): http.Client.Do can call payload.Read, reading the content of payload.buf, while returning err != nil, thus leading to the payload not being reset, and a state where payload.buf is empty, while count !=0, since pushing a trace to the payload doesn't reset payload.count to 0, and the msgp array header depends on payload.count, this results in subsequent payload Reads returning an incorrect msgpack serialization of the traces array. Subsequent flushes to the agent therefore always return an error (Status: Bad Request), since the agent is unable to decode the incorrectly serialized traces array, while  payload.buf's content is still being read, thus repeating the cycle. The only way to get out of this corrupt state by then is for the service to generate enough traces to make the payload exceed payloadMaxSize before the next flush. 
This is rather bothersome because it makes the instrumented service stop reporting APM traces, as traces are lost (being read and therefore lost on the tracer end, while not being decodable by the DataDog agent).

This issue typically happens when the DataDog agent is restarted, or when deploying a new version of the agent: all our production Ruby services keep on having their APM traces collected, while some go services don't, depending on whether they were flushing traces at the moment the newly started agent began listening on port 8126 (and read the "traces flush" request body without being able to process it, returning an error while having emptied the payload buffer on the instrumented services' end).

This pull request solves the issue by modifying the payload to make it impossible to have a mismatch between the msgpack header and the actual traces content of the payload. 
I also modified the payload benchmark to be more meaningful by actually measuring the whole payload lifecycle: pushing traces to a payload, then reading the whole payload, instead of only measuring pushing traces to the payload.

The new payload implementation performs the same as the former, while allocating much less memory for large payload sizes (10 MB).

Benchmark results on my Machine: 

- Previous implementation: 
`goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
BenchmarkPayloadThroughput/100KB-8         	  200000	     10198 ns/op	     321 B/op	      10 allocs/op
BenchmarkPayloadThroughput/1MB-8           	   10000	    109813 ns/op	     712 B/op	      10 allocs/op
BenchmarkPayloadThroughput/10MB-8          	    1000	   1618488 ns/op	   32120 B/op	      10 allocs/op
PASS
ok  	gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer	5.124s`

- New implementation from this PR: 
`goos: darwin
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
BenchmarkPayloadThroughput/100KB-8         	  200000	     10786 ns/op	    1073 B/op	      15 allocs/op
BenchmarkPayloadThroughput/1MB-8           	   10000	    117055 ns/op	    1102 B/op	      15 allocs/op
BenchmarkPayloadThroughput/10MB-8          	    1000	   1300400 ns/op	    3752 B/op	      15 allocs/op
PASS
ok  	gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer	4.934s`

I hope my explanation was clear enough, and am available if you require any additional information.