@gbbr 2018-07-20
Indeed you are right, it is not possible to read multiple times from a `bytes.Buffer`. This is an unfortunate omission on our side. There are however multiple solutions to this issue that we can discuss.

Thanks a lot from going through the effort of attempting to write a solution to this. I'd like to kindly ask you to forgive us for possibly not being obvious enough with our [contribution guidelines](https://github.com/DataDog/dd-trace-go/blob/v1/CONTRIBUTING.md), but I am going to have to close this pull request. Please take a moment to read them.

It is super ok to submit very obvious small fixes or typos, but for core changes such as this one they definitely need to be discussed and concluded on in an issue before submitting to writing any actual code. I think there are several solutions we can consider to this problem. I've temporarily opened an issue [here](https://github.com/DataDog/dd-trace-go/issues/276) where we can discuss any ideas and solutions if you are open to it. I'd also be happy for you to open the issue yourself so you can keep track of its progress (in which case I will close mine).

Once we reach a consensus, I'd be more than happy for you to start hacking on it. This is definitely something we are interested in fixing.