@shidenkai0 2018-07-20
Hi,
My misunderstanding, I thought this entered in the scope of bufixes, since it solves a bug in an underlying type. I didn't think it might be considered a change in functionality or features. Will be glad to discuss on the associated issue.

Thanks,