@gbbr 2018-07-20
It's all good. We could've done a better job at writing those guidelines. I'll make some changes there to be more explicit 👍