@shidenkai0 2018-07-20
Regarding the payload benchmark, I compared by rolling back payload.go to the previous version of the file, while using the same benchmark, so I think it is still relevant to compare both implementations. I also think that benchmarking the whole payload lifecycle (push traces and read them) better represents the actual impact of the payload on a running service process.  I would be glad to discuss this choice.