@shidenkai0 2018-07-20
Totally, do you want me to make separate benchmarks for payload filling, and reading ?