@gbbr 2018-07-20
If you don't mind doing so, I think they'd both deem to be very beneficial. For payload filling, I'd prefer to keep the one we originally had unless there's a good reason for it to change.