@shidenkai0 2018-07-20
I reverted to the original payload throughput benchmark, and implemented a payload Read benchmark. I will be awaiting your review.