@gbbr 2018-07-22
Nice! Thanks for adding the benchmark. It will definitely come in handy. Would you mind posting the results for both of your benchmarks, now that you've put back the original one?