@gbbr 2018-07-22
I like this implementation because it significantly speeds up the throughput and moves the encoding into the read part. Eventually we might want to make the flushing it's own go routine and then blocking will be minimal.

Before moving further, it would be nice to also have a test which reproduces the problem that we are fixing.