@gbbr 2018-07-22
I hope you don't mind all my comments today. Since I was bored on Sunday, I ran the benchmarks for you out of curiosity. Here are the results:
```go
name                      old time/op    new time/op    delta
PayloadThroughput/10K-4      738µs ± 0%      48µs ± 0%   -93.53%
PayloadThroughput/100K-4     676µs ± 0%      40µs ± 0%   -94.06%
PayloadThroughput/1MB-4      705µs ± 0%      40µs ± 0%   -94.29%
PayloadRead/10K-4           4.28ms ± 0%    0.79ms ± 0%   -81.51%
PayloadRead/100K-4          4.27ms ± 0%    0.80ms ± 0%   -81.17%
PayloadRead/1MB-4           4.38ms ± 0%    1.28ms ± 0%   -70.82%

name                      old alloc/op   new alloc/op   delta
PayloadThroughput/10K-4     46.7kB ± 0%     0.0kB ± 0%  -100.00%
PayloadThroughput/100K-4    19.0kB ± 0%     0.0kB       -100.00%
PayloadThroughput/1MB-4     16.2kB ± 0%     0.0kB       -100.00%
PayloadRead/10K-4           31.9MB ± 0%     0.1MB ± 0%   -99.54%
PayloadRead/100K-4          31.9MB ± 0%     0.3MB ± 0%   -99.02%
PayloadRead/1MB-4           31.9MB ± 0%     2.6MB ± 0%   -91.77%

name                      old allocs/op  new allocs/op  delta
PayloadThroughput/10K-4        962 ± 0%         0       -100.00%
PayloadThroughput/100K-4      97.0 ± 0%       0.0       -100.00%
PayloadThroughput/1MB-4       10.0 ± 0%       0.0       -100.00%
PayloadRead/10K-4              990 ± 0%       982 ± 0%    -0.81%
PayloadRead/100K-4             122 ± 0%       121 ± 0%    -0.82%
PayloadRead/1MB-4             34.0 ± 0%      37.0 ± 0%    +8.82%
```
I think this is fantastic! Everything's almost doubled and is now more correct. I'm definitely happy to merge this once we bring it into its final state.

Thanks once again.