@gbbr 2018-07-25
Fixes #276