@dd-caleb 2018-07-25
As an experiment I tried using an io.Pipe to create a ReadCloser which simplifies the code quite a bit:

```go
// ReadCloser returns an io.ReadCloser. It reads from the msgpack-encoded stream.
func (p *payload) ReadCloser() io.ReadCloser {
	pr, pw := io.Pipe()
	go func() {
		bpw := bufio.NewWriter(pw)

		_, err := bpw.Write(p.header[p.off:])
		if err != nil {
			pw.CloseWithError(err)
			return
		}

		for _, trace := range p.traces {
			err = msgp.Encode(bpw, trace)
			if err != nil {
				pw.CloseWithError(err)
				return
			}
		}

		if err := bpw.Flush(); err != nil {
			pw.CloseWithError(err)
		} else {
			pw.Close()
		}
	}()
	return pr
}
```

Unfortunately it slows down the reads quite a bit:

```
BenchmarkPayloadThroughput/10K-4      59444         61024          +2.66%
BenchmarkPayloadThroughput/100K-4     50649         54035          +6.69%
BenchmarkPayloadThroughput/1MB-4      51219         54499          +6.40%
BenchmarkPayloadRead/10K-4            1136991       7363187        +547.60%
BenchmarkPayloadRead/100K-4           1228897       6874977        +459.44%
BenchmarkPayloadRead/1MB-4            2097430       7064675        +236.83%
```

Though I wonder if we might be able to implement something similar without the `io.Pipe`. I will do another experiment and post my results.