@dd-caleb 2018-07-25
OK this version is roughly the same in terms of performance:

```go

// Reader returns an io.Reader. It reads from the msgpack-encoded stream.
func (p *payload) Reader() io.Reader {
	pr := &payloadReader{
		traces: p.traces,
	}
	// this never returns an error which is why its ignored
	pr.buf.Write(p.header[p.off:])
	return pr
}

type payloadReader struct {
	traces []spanList
	buf    bytes.Buffer
}

func (pr *payloadReader) Read(b []byte) (n int, err error) {
	for len(pr.traces) > 0 && pr.buf.Len() < len(b) {
		err := msgp.Encode(&pr.buf, pr.traces[0])
		if err != nil {
			return 0, err
		}
		pr.traces = pr.traces[1:]
	}
	return pr.buf.Read(b)
}
```

```
BenchmarkPayloadThroughput/10K-4      59444         63686         +7.14%
BenchmarkPayloadThroughput/100K-4     50649         52404         +3.47%
BenchmarkPayloadThroughput/1MB-4      51219         52836         +3.16%
BenchmarkPayloadRead/10K-4            1136991       1039719       -8.56%
BenchmarkPayloadRead/100K-4           1228897       1187651       -3.36%
BenchmarkPayloadRead/1MB-4            2097430       2095003       -0.12%
```

Feel free to incorporate this change if you feel like it makes the code cleaner. I'm fine with it either way.